backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
diff-match-patch==20241021
Django==5.2.7
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.9
django-js-asset==3.1.2
django-parler==2.3
django-parler-rest==2.2
django-redis==6.0.0
django-unfold==0.65.0
djangorestframework==3.16.1
dnspython==2.8.0
idna==3.11
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
python-dotenv==1.1.1
redis==7.0.0
requests==2.32.5
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
urllib3==2.5.0
